chore: handle group room reservations
