+++ frontend/src/app/coworking/coworking.module.ts
@@ -14,16 +14,11 @@ import { ReservationComponent } from './reservation/reservation.component';
 import { MatIconModule } from '@angular/material/icon';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatButtonModule } from '@angular/material/button';
-import { NewReservationPageComponent } from './room-reservation/new-reservation-page/new-reservation-page.component';
-import { RoomReservationWidgetComponent } from './widgets/room-reservation-table/room-reservation-table.widget';
-import { ConfirmReservationComponent } from './room-reservation/confirm-reservation/confirm-reservation.component';
-import { DateSelector } from './widgets/date-selector/date-selector.widget';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatNativeDateModule } from '@angular/material/core';
 import { OperatingHoursDialog } from './widgets/operating-hours-dialog/operating-hours-dialog.widget';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
-import { ReactiveFormsModule } from '@angular/forms';
 import { MatInputModule } from '@angular/material/input';
 import { SharedModule } from '../shared/shared.module';
 import { MatTabsModule } from '@angular/material/tabs';
@@ -31,14 +26,11 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { AmbassadorXLComponent } from './ambassador-home/ambassador-xl/ambassador-xl.component';
 import { AmbassadorRoomComponent } from './ambassador-home/ambassador-room/ambassador-room.component';
 import { ReservationFactsWidget } from './widgets/reservation-facts/reservation-facts.widget';
-import { DialogModule } from '@angular/cdk/dialog';
 import { MatDialogModule } from '@angular/material/dialog';
-import { NewRoomReservationComponent } from './new-room-reservation/new-room-reservation.component';
+import { NewRoomReservationComponent } from './room-reservation/room-reservation.component';
 
 @NgModule({
   declarations: [
-    NewReservationPageComponent,
-    RoomReservationWidgetComponent,
     CoworkingPageComponent,
     ReservationComponent,
     AmbassadorPageComponent,
@@ -46,9 +38,6 @@ import { NewRoomReservationComponent } from './new-room-reservation/new-room-res
     AmbassadorRoomComponent,
     CoworkingDropInCard,
     CoworkingReservationCard,
-    ConfirmReservationComponent,
-    NewReservationPageComponent,
-    DateSelector,
     OperatingHoursDialog,
     ReservationFactsWidget,
     NewRoomReservationComponent
