--- frontend/src/app/coworking/room-reservation/confirm-reservation/confirm-reservation.component.css
@@ -1,34 +1,0 @@
-.card-container {
-  margin: 16px;
-}
-
-.mat-mdc-card {
-  max-width: 640px;
-  margin: 0px;
-}
-
-.mat-mdc-card-header {
-  margin-bottom: 16px;
-}
-
-.mat-mdc-card-actions {
-  margin-top: 16px;
-}
-
-.mat-divider {
-  margin: 1em 0;
-}
-
-h3 {
-  font-size: 18px;
-  margin-bottom: 0px;
-}
-
-h3 > label {
-  width: 64px;
-  display: inline-block;
-}
-
-p {
-  margin-left: 64px;
-}
