--- frontend/src/app/coworking/room-reservation/confirm-reservation/confirm-reservation.component.html
@@ -1,8 +1,0 @@
-<div class="card-container">
-  @if(reservation) {
-  <coworking-reservation-card
-    [reservation]="reservation"
-    (isConfirmed)="setConfirmation($event)">
-  </coworking-reservation-card>
-  }
-</div>
