--- frontend/src/app/coworking/room-reservation/new-reservation-page/new-reservation-page.component.css
@@ -1,78 +1,0 @@
-::ng-deep .mat-mdc-card-outlined {
-  max-width: 100% !important;
-  margin-right: 32px !important;
-}
-
-.mat-mdc-card-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 12px;
-}
-
-.mat-mdc-card-title {
-  padding-left: 10px;
-}
-
-.mat-mdc-card-subtitle {
-  padding-left: 10px;
-}
-
-
-/* Add any additional styling for the room-reservation-table if needed */
-
-.date-selector {
-  padding: 10px; /* Adjust the value as needed */
-}
-
-.upcoming-reservations-container {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px; /* Adjust the gap as needed */
-}
-
-.coworking-reservation-card {
-  width: calc(50% - 10px); /* Adjust the width and margin as needed */
-  margin-bottom: 20px;
-  box-sizing: border-box; /* Include padding and border in the total width */
-}
-
-.heading {
-  padding-top: 30px;
-  padding-left: 10px;
-}
-
-.legend-container {
-  display: flex;
-  flex-wrap: wrap;
-  margin-top: 20px; /* Adjust the margin as needed */
-}
-
-.legend-item {
-  display: flex;
-  align-items: center;
-  margin-left: 10px; /* Adjust the margin between legend items */
-  margin-top: 5px;
-}
-
-.legend-color {
-  width: 20px;
-  height: 20px;
-  margin-right: 5px;
-  border-radius: 6px;
-}
-
-.legend-text {
-  font-size: 14px;
-}
-
-.reservation-limit {
-  padding-left: 10px;
-}
-
-@media only screen and (max-width: 900px) {
-  .mat-mdc-card-header {
-    flex-direction: column;
-    align-items: baseline;
-  }
-}
