+++ frontend/src/app/coworking/room-reservation/room-reservation.component.html
@@ -2,7 +2,7 @@
     <mat-card-header class="flex flex-row items-center justify-between p-4">
         <mat-card-title>Reserve a Room</mat-card-title>
         <div class="date-selector">
-            <date-selector></date-selector>
+            <!-- <date-selector></date-selector> -->
         </div>
     </mat-card-header>
