--- frontend/src/app/coworking/widgets/date-selector/date-selector.widget.html
@@ -1,11 +1,0 @@
-<mat-form-field appearance="outline" subscriptSizing="dynamic">
-  <input
-    matInput
-    [matDatepicker]="picker"
-    [value]="minDate"
-    [min]="minDate"
-    [max]="maxDate"
-    (dateChange)="onDateChange($event)" />
-  <mat-datepicker-toggle matSuffix [for]="picker"></mat-datepicker-toggle>
-  <mat-datepicker #picker></mat-datepicker>
-</mat-form-field>
