--- frontend/src/app/coworking/widgets/room-reservation-table/room-reservation-table.widget.html
@@ -1,36 +1,0 @@
-<div class="table-container">
-  <div class="table-responsive">
-    <table class="table">
-      <tr>
-        <th></th>
-        @for(timeSlot of timeSlots; track timeSlot) {
-        <th [innerHTML]="timeSlot" class="time-slot"></th>
-        }
-      </tr>
-      @for (record of reservationsMap | keyvalue; track record.key) {
-      <tr>
-        <th>{{ record.key }}
-          <br><span class= "room-details"> {{"# seats: "}}{{capacityMap[record.key]}}
-          <br>{{roomTypeMap[record.key]}}</span>
-        </th>
-        <td
-          *ngFor="let timeSlot of record.value; let j = index"
-          [style.background]="cellPropertyMap[timeSlot].backgroundColor"
-          (click)="
-            !cellPropertyMap[timeSlot].isDisabled &&
-              toggleCellColor(record.key, j)
-          "></td>
-      </tr>
-      }
-    </table>
-  </div>
-</div>
-<div class="divider">
-  <button
-    mat-stroked-button
-    class="button"
-    (click)="draftReservation()"
-    [disabled]="!selectButton">
-    Reserve
-  </button>
-</div>
